+++ components/modals/SelectedPlayerModal.tsx
@@ -5,7 +5,7 @@ import { Image } from "expo-image";
 import { MaterialCommunityIcons, FontAwesome5, Ionicons } from "@expo/vector-icons";
 import { Text, Separator } from "@/components/Themed";
 import openExternalLink from "@/utils/openExternalLink";
-import ServerPlayer from "@/interfaces/ServerPlayer";
+import ServerPlayer from "@/types/ServerPlayer";
 
 type SelectedPlayerModalProps = {
