+++ types/ConsoleOutput.ts
