+++ types/ServerBan.ts
