+++ types/ServerInfo.ts
